Move wow64 process check into initialize
